app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
pyasn1==0.4.8
pymongo==4.0.2
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
Telethon==1.24.0
urllib3==1.26.9
